turns Type into a URL link
